requirements.txt
attrs==21.2.0
backcall==0.2.0
click==8.0.3
decorator==5.1.0
environs==9.3.4
Flask==2.0.2
iniconfig==1.1.1
ipython==7.28.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.2
MarkupSafe==2.0.1
marshmallow==3.14.0
matplotlib-inline==0.1.3
packaging==21.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
pluggy==1.0.0
prompt-toolkit==3.0.20
psycopg2-binary==2.9.1
ptyprocess==0.7.0
py==1.10.0
Pygments==2.10.0
pyparsing==2.4.7
python-dotenv==0.19.1
toml==0.10.2
traitlets==5.1.0
wcwidth==0.2.5
Werkzeug==2.0.2
